+++ apps/docs/next.config.mjs
@@ -23,6 +23,13 @@ const config = {
 };
 
 const withMDX = createMDX({
+  buildSearchIndex: {
+    filter: (v) => {
+      if (v.match(/.+\.model\.mdx/)) return false;
+
+      return true;
+    },
+  },
   mdxOptions: {
     rehypeCodeOptions: {
       transformers: [