+++ packages/core/src/search/client.ts
@@ -31,15 +31,22 @@ async function fetchDocs(
   return (await res.json()) as SortedResult[];
 }
 
-export function useDocsSearch(locale?: string, tag?: string): UseDocsSearch {
+/**
+ * @param locale - Filter with locale
+ * @param tag - Filter with specific tag
+ * @param api - The Search API URL
+ */
+export function useDocsSearch(
+  locale?: string,
+  tag?: string,
+  api = '/api/search',
+): UseDocsSearch {
   const [search, setSearch] = useState('');
   const debouncedValue = useDebounce(search, 100);
 
   const query = useSWR(
-    ['/api/search', debouncedValue, locale, tag],
-    async ([api, value]) => {
-      return fetchDocs(api, value, locale, tag);
-    },
+    [api, debouncedValue, locale, tag],
+    (args) => fetchDocs(...args),
     {
       keepPreviousData: true,
     },