+++ packages/core/src/source/index.ts
@@ -1,5 +1,7 @@
 export * from './page-tree-builder';
-export * from './types';
 export * from './loader';
 export * as FileSystem from './file-system';
-export * from './load';
+
+export type { LoadOptions, Transformer, VirtualFile } from './load';
+export type * from './types';
+export { type FileInfo, parseFilePath, parseFolderPath } from './path';