+++ packages/mdx/src/loader-mdx.ts
@@ -13,8 +13,8 @@ export interface Options extends ProcessorOptions {
   lastModifiedTime?: 'git' | 'none';
 }
 
-interface InternalBuildInfo {
-  __next_docs?: {
+export interface InternalBuildInfo {
+  __fumadocs?: {
     path: string;
     /**
      * `vfile.data` parsed from file
@@ -71,12 +71,12 @@ export default async function loader(
       (file) => {
         const info = this._module?.buildInfo as InternalBuildInfo;
 
-        info.__next_docs = {
+        info.__fumadocs = {
           path: filePath,
           data: file.data,
         };
 
-        callback(undefined, String(file.value), file.map || undefined);
+        callback(undefined, String(file.value), file.map ?? undefined);
       },
       (error: Error) => {
         const fpath = path.relative(context, filePath);