+++ packages/typescript/test/fixtures/test-2.ts
@@ -19,4 +19,9 @@ export interface Player {
    * Returned by API
    */
   age: number;
+
+  /**
+   * @internal
+   */
+  privateValue: string
 }