+++ packages/ui/src/components/dialog/search-default.tsx
@@ -9,12 +9,13 @@ export type DefaultSearchDialogProps = SharedProps & ContentProps;
 
 export default function DefaultSearchDialog({
   tag,
+  api,
   footer,
   ...props
 }: DefaultSearchDialogProps): JSX.Element {
   return (
     <SearchDialog {...props}>
-      <Content tag={tag} footer={footer} />
+      <Content tag={tag} footer={footer} api={api} />
     </SearchDialog>
   );
 }
@@ -25,12 +26,17 @@ interface ContentProps {
    */
   tag?: string;
 
+  /**
+   * Search API URL
+   */
+  api?: string;
+
   footer?: ReactNode;
 }
 
-function Content({ tag, footer }: ContentProps): JSX.Element {
+function Content({ tag, footer, api }: ContentProps): JSX.Element {
   const { locale } = useI18n();
-  const { search, setSearch, query } = useDocsSearch(locale, tag);
+  const { search, setSearch, query } = useDocsSearch(locale, tag, api);
 
   return (
     <SearchDialogContent