+++ packages/ui/src/components/ui/scroll-area.tsx
@@ -11,7 +11,7 @@ const ScrollArea = React.forwardRef<
     className={cn('overflow-hidden', className)}
     {...props}
   >
-    <ScrollAreaPrimitive.Viewport className="h-full w-full rounded-[inherit]">
+    <ScrollAreaPrimitive.Viewport className="size-full rounded-[inherit]">
       {children}
     </ScrollAreaPrimitive.Viewport>
     <ScrollAreaPrimitive.Corner />