+++ packages/contentlayer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # fumadocs-contentlayer
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [c9b7763]
+- Updated dependencies [0e78dc8]
+- Updated dependencies [d8483a8]
+  - fumadocs-core@10.0.1
+
 ## 1.1.1
 
 ### Patch Changes